+++ src/main/java/dashboard/OperatorController.java
@@ -1,4 +1,4 @@
-package com.uniovi.controllers;
+package dashboard;
 
 import java.security.Principal;
 import java.util.List;
