+++ src/test/java/test/controller/HomeControllerTest.java
@@ -13,7 +13,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
-import com.uniovi.controllers.HomeController;
+import dashboard.HomeController;
 
 public class HomeControllerTest {
