+++ src/test/java/test/controller/OperatorsControllerTest.java
@@ -16,11 +16,12 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
-import com.uniovi.controllers.OperatorController;
 import com.uniovi.entities.Incident;
 import com.uniovi.main.InciDashboardI2bApplication;
 import com.uniovi.services.IncidentsService;
 import com.uniovi.services.OperatorsService;
+
+import dashboard.OperatorController;
 
 @SpringBootTest(classes = { InciDashboardI2bApplication.class })
 @RunWith(SpringJUnit4ClassRunner.class)
